Initial commit for project restructure
