+++ lib/services/claims/infra/tests/claims-service.test.js
@@ -1,0 +1,1 @@
+// Infrastructure test
