+++ lib/services/documents/app/handlers/analyzeCarImage.js
