+++ lib/services/documents/app/handlers/textractResponseTransformer.js
