+++ package.json
@@ -38,7 +38,7 @@
   },
   "dependencies": {
     "@aws-lambda-powertools/metrics": "^1.5.0",
-    "aws-cdk-lib": "^2.60.0",
+    "aws-cdk-lib": "^2.67.0",
     "cdk-nag": "^2.21.58",
     "constructs": "^10.1.216",
     "source-map-support": "^0.5.21",
